# docs/swagger.yaml
openapi: 3.0.3
info:
  title: PTEST
  version: 1.0.0
  description: |
    API for managing property agents, rental properties, family-based tenants,
    property notes, and operational reminders.

    Core rules:
    1. One agent can manage multiple properties.
    2. Each property can have only one family household.
    3. A family can contain multiple tenants.
    4. Notes and reminders are authored by an agent for a property.
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      security: []
      tags: [Auth]
      summary: Register a user
      description: Creates a new auth user and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /auth/login:
    post:
      security: []
      tags: [Auth]
      summary: Login user
      description: Authenticates a user and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: Returns the currently authenticated user from the bearer token.
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
...
    UnauthorizedError:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    ConflictError:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter only the JWT token value (without the `Bearer ` prefix).
  schemas:
    AuthUser:
      type: object
      properties:
        id: { type: integer, example: 1 }
        fullName: { type: string, example: Admin User }
        email: { type: string, format: email, example: admin@pureagent.local }
      required: [id, fullName, email]
    AuthResponse:
      type: object
      properties:
        token:
          { type: string, example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9... }
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [token, user]
    RegisterRequest:
      type: object
      properties:
        fullName: { type: string, minLength: 1 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
      required: [fullName, email, password]
    LoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
      required: [email, password]
    ErrorMessage:
      type: object
      properties:
        error:
          type: string
      required: [error]
...